+++ src/renderer/src/components/ModelLibrary/providers/index.ts
@@ -6,6 +6,7 @@ export * from './DeepseekParameters';
 export * from './GeminiParameters';
 export * from './GpustackParameters';
 export * from './GroqParameters';
+export * from './MinimaxParameters';
 export * from './LmStudioParameters';
 export * from './OllamaParameters';
 export * from './OpenAiParameters';